GENESIS COMMIT!
[ADD]
=> Pushed all the work done in local between various computer to finally a clean repo.
Now this shit is getting serious.

Advancement so far:
Global structure for v1.0 is completed at 90%, only need to add clarity to the code with comments and rigor and coding style
Bugs needs to be fixed, primordial bugs for CE tickets and major bugs for ATY, they are all in the "problems to solve.tct" file.
We will need to be more percise in the future about them.

This will be the fundamentals about our future commit-norm:
"version" [ADD] => adding new functions / files / functionnality
"version" [MOD] => modifiyng functions or method
"version" [SUP] => deleting methods or functions

Hopefully, the "Site/" folder will soon be beefier :)